extract post published action into own file
