+++ src/handlers.ts
@@ -1,28 +1,25 @@
 import { createHash } from 'node:crypto';
 import {
     type Actor,
-    Article,
     Create,
     Follow,
     Like,
     Mention,
     Note,
     PUBLIC_COLLECTION,
-    type RequestContext,
     Undo,
     isActor,
 } from '@fedify/fedify';
 import { Temporal } from '@js-temporal/polyfill';
-import type { Context, Next } from 'hono';
+import type { Context } from 'hono';
 import { v4 as uuidv4 } from 'uuid';
-import { type ContextData, type HonoContextVariables, fedify } from './app';
+import { type HonoContextVariables, fedify } from './app';
 import { ACTOR_DEFAULT_HANDLE } from './constants';
 import {
     buildActivity,
     prepareNoteContent,
 } from './helpers/activitypub/activity';
 import { escapeHtml } from './helpers/html';
-import { toURL } from './helpers/uri';
 import { getUserData } from './helpers/user';
 import { addToList, removeFromList } from './kv-helpers';
 import { lookupActor, lookupObject } from './lookup-helpers';
@@ -31,60 +28,6 @@ import z from 'zod';
 import { getSite } from './db';
 import { updateSiteActor } from './helpers/activitypub/actor';
 import { getSiteSettings } from './helpers/ghost';
-
-enum PostVisibility {
-    Public = 'public',
-    Members = 'members',
-    Paid = 'paid',
-    Tiers = 'tiers',
-}
-
-const PostSchema = z.object({
-    uuid: z.string().uuid(),
-    title: z.string(),
-    html: z.string().nullable(),
-    excerpt: z.string().nullable(),
-    feature_image: z.string().url().nullable(),
-    published_at: z.string().datetime(),
-    url: z.string().url(),
-    visibility: z.nativeEnum(PostVisibility),
-});
-
-type Post = z.infer<typeof PostSchema>;
-
-async function postToArticle(
-    ctx: RequestContext<ContextData>,
-    post: Post,
-    author: Actor | null,
-) {
-    if (!post) {
-        return {
-            article: null,
-            preview: null,
-        };
-    }
-    const preview = new Note({
-        id: ctx.getObjectUri(Note, { id: post.uuid }),
-        content: post.excerpt,
-    });
-    const article = new Article({
-        id: ctx.getObjectUri(Article, { id: post.uuid }),
-        attribution: author,
-        name: post.title,
-        content: post.html,
-        image: toURL(post.feature_image),
-        published: Temporal.Instant.from(post.published_at),
-        preview: preview,
-        url: toURL(post.url),
-        to: PUBLIC_COLLECTION,
-        cc: ctx.getFollowersUri(ACTOR_DEFAULT_HANDLE),
-    });
-
-    return {
-        article,
-        preview,
-    };
-}
 
 export async function unlikeAction(
     ctx: Context<{ Variables: HonoContextVariables }>,
@@ -492,85 +435,6 @@ export async function followAction(
     );
     // We return the actor because the serialisation of the object property is not working as expected
     return new Response(JSON.stringify(await actorToFollow.toJsonLd()), {
-        headers: {
-            'Content-Type': 'application/activity+json',
-        },
-        status: 200,
-    });
-}
-
-const PostPublishedWebhookSchema = z.object({
-    post: z.object({
-        current: PostSchema,
-    }),
-});
-
-export async function postPublishedWebhook(
-    ctx: Context<{ Variables: HonoContextVariables }>,
-    next: Next,
-) {
-    const logger = ctx.get('logger');
-    const data = PostPublishedWebhookSchema.parse(
-        (await ctx.req.json()) as unknown,
-    );
-
-    if (data.post.current.visibility !== PostVisibility.Public) {
-        logger.info('Post is not public, skipping');
-
-        return new Response(JSON.stringify({}), {
-            headers: {
-                'Content-Type': 'application/activity+json',
-            },
-            status: 200,
-        });
-    }
-
-    const apCtx = fedify.createContext(ctx.req.raw as Request, {
-        db: ctx.get('db'),
-        globaldb: ctx.get('globaldb'),
-        logger: ctx.get('logger'),
-    });
-    const actor = await apCtx.getActor(ACTOR_DEFAULT_HANDLE);
-    const { article, preview } = await postToArticle(
-        apCtx,
-        data.post.current,
-        actor,
-    );
-    if (article) {
-        const create = new Create({
-            actor,
-            object: article,
-            id: apCtx.getObjectUri(Create, { id: uuidv4() }),
-            to: PUBLIC_COLLECTION,
-            cc: apCtx.getFollowersUri('index'),
-        });
-        try {
-            await article.toJsonLd();
-            await ctx
-                .get('globaldb')
-                .set([preview.id!.href], await preview.toJsonLd());
-            await ctx
-                .get('globaldb')
-                .set([create.id!.href], await create.toJsonLd());
-            await ctx
-                .get('globaldb')
-                .set([article.id!.href], await article.toJsonLd());
-            await addToList(ctx.get('db'), ['outbox'], create.id!.href);
-            await apCtx.sendActivity(
-                { handle: ACTOR_DEFAULT_HANDLE },
-                'followers',
-                create,
-                {
-                    preferSharedInbox: true,
-                },
-            );
-        } catch (err) {
-            ctx.get('logger').error('Post published webhook failed: {error}', {
-                error: err,
-            });
-        }
-    }
-    return new Response(JSON.stringify({}), {
         headers: {
             'Content-Type': 'application/activity+json',
         },
